+++ egui/src/input.rs
@@ -151,7 +151,7 @@ impl Default for MouseInput {
 }
 
 /// An input event. Only covers events used by Egui.
-#[derive(Clone, Debug, Eq, Ord, PartialEq, PartialOrd)]
+#[derive(Clone, Debug, Eq, PartialEq)]
 pub enum Event {
     Copy,
     Cut,
@@ -161,33 +161,45 @@ pub enum Event {
     Key {
         key: Key,
         pressed: bool,
+        modifiers: Modifiers,
     },
 }
 
-/// Keyboard key name. Only covers keys used by Egui.
+/// State of the modifier keys. These must be fed to Egui.
+#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
+pub struct Modifiers {
+    /// Either of the alt keys are down (option ⌥ on Mac)
+    pub alt: bool,
+    /// Either of the control keys are down
+    pub ctrl: bool,
+    /// Either of the shift keys are down
+    pub shift: bool,
+    /// The Mac ⌘ Command key. Should always be set to `false` on other platforms.
+    pub mac_cmd: bool,
+    /// On Mac, this should be set whenever one of the ⌘ Command keys are down (same as `mac_cmd`).
+    /// On Windows and Linux, set this to the same value as `ctrl`.
+    /// This is so that Egui can, for instance, select all text by checking for `command + A`
+    /// and it will work on both Mac and Windows.
+    pub command: bool,
+}
+
+/// Keyboard key name. Only covers keys used by Egui (mostly for text editing).
 #[derive(Clone, Copy, Debug, Eq, Ord, PartialEq, PartialOrd)]
 pub enum Key {
-    Alt,
+    ArrowDown,
+    ArrowLeft,
+    ArrowRight,
+    ArrowUp,
     Backspace,
-    Control,
     Delete,
-    Down,
     End,
+    Enter,
     Escape,
     Home,
     Insert,
-    Left,
-    /// Windows key or Mac Command key
-    Logo,
     PageDown,
     PageUp,
-    /// Enter/Return key
-    Enter,
-    Right,
-    Shift,
-    // Space,
     Tab,
-    Up,
 }
 
 impl InputState {
@@ -227,7 +239,8 @@ impl InputState {
                 event,
                 Event::Key {
                     key,
-                    pressed: true
+                    pressed: true,
+                    ..
                 } if *key == desired_key
             )
         })
@@ -240,7 +253,8 @@ impl InputState {
                 event,
                 Event::Key {
                     key,
-                    pressed: false
+                    pressed: false,
+                    ..
                 } if *key == desired_key
             )
         })